Компонент для списка треков в плеере
