+++ static/css/partials/_player.scss
@@ -244,7 +244,7 @@
 
 .track-list .track-name {
   font-size: 18px;
-  width: 276px;
+  width: 230px;
 }
 
 .track-list .group-name {
