+++ static/js/components/player_track_list_component/player_track_list_component.js
@@ -1,0 +1,196 @@
+import {PLAYER, GLOBAL, URL} from '@libs/constans';
+import template from '@components/player_track_list_component/player_track_list_component.tmpl.xml';
+import User from '@libs/user';
+
+/**
+ * Компонент - список треков в плеере
+ */
+export default class TrackListComponent {
+    /**
+     * Конструткор
+     * @param {Object} eventBus
+     */
+    constructor(eventBus) {
+        this.eventBus = eventBus;
+    }
+
+    /**
+     * Sets EventListeners
+     */
+    setEventListeners() {
+        console.log(document.querySelectorAll('.favorite-button').length);
+        document.querySelectorAll('.track-list').forEach((row) => {
+            row.onclick = (event) => this.tracklistClick(event);
+        });
+        document.querySelectorAll('.delete-button').forEach((button) => {
+            button.onclick = (event) => this.trackDeleteButtonClick(event);
+        });
+        document.querySelectorAll('.favorite-button').forEach((button) => {
+            button.onclick = (event) => this.trackFavoriteButtonClick(event);
+        });
+        document.querySelectorAll('.add-button').forEach((button) => {
+            button.onclick = (event) => this.trackAddButtonClick(event);
+        });
+    }
+
+    /**
+     * Слушает скрол для прокрутки плейлиста
+     * @param {Object} event
+     */
+    trackListWheel(event) {
+        const delta = event.deltaY;
+        const trackList = document.getElementsByClassName('track-list')[0];
+        if (event.clientX > trackList.getBoundingClientRect().x &&
+            event.clientX < trackList.getBoundingClientRect().x +
+            trackList.getBoundingClientRect().width &&
+            event.clientY > trackList.getBoundingClientRect().y &&
+            event.clientY < trackList.getBoundingClientRect().y +
+            trackList.getBoundingClientRect().height
+        ) {
+            event.preventDefault();
+            const top = parseInt(trackList.style.top.slice(0, trackList.style.top.length - 2));
+            if (delta < 0 && top < 0 ||
+                delta > 0 && trackList.getBoundingClientRect().bottom >
+                document.documentElement.clientHeight
+            ) {
+                if (delta < 0 && top - delta / 2 > 0) {
+                    trackList.style.top = '0';
+                } else if (delta > 0 && trackList.getBoundingClientRect().bottom - delta / 2 <
+                    document.documentElement.clientHeight
+                ) {
+                    const container = document.getElementsByClassName('container-audio')[0];
+                    trackList.style.top = (document.documentElement.clientHeight -
+                        trackList.getBoundingClientRect().height -
+                        container.getBoundingClientRect().bottom).toString() + 'px';
+                } else {
+                    trackList.style.top = (top - delta / 2).toString() + 'px';
+                }
+            }
+        }
+    }
+
+    /**
+     * Слушает клик по треку в плейлисте
+     * @param {Object} event
+     */
+    tracklistClick(event) {
+        let current = event.target;
+        if (current.classList.contains('m-obscure-title') ||
+            current.classList.contains('m-big-delete-button')) {
+            return;
+        }
+        while (current !== window && current !== document.body && current != null) { // TODO Сделать по-человечески :(
+            if (current.getAttribute('class') === 'track-list' ||
+                (current.getAttribute('class') !== null &&
+                    current.getAttribute('class').indexOf('button') !== -1 &&
+                    current.getAttribute('class').indexOf('buttons') === -1 &&
+                    current.getAttribute('class').indexOf('row') !== -1)
+            ) {
+                break;
+            }
+            if (current.getAttribute('id') !== null) {
+                this.eventBus.emit(PLAYER.GET_TRACK, current.getAttribute('id'));
+                break;
+            } else {
+                current = current.parentNode;
+            }
+        }
+    }
+
+    /**
+     * Слушает клик мыши по кнопке удаления на треке в плейлисте
+     * @param {Object} event
+     */
+    trackDeleteButtonClick(event) {
+        let target = event.target;
+        while (target.getAttribute('id') === null) {
+            target = target.parentNode;
+        }
+        this.eventBus.emit(PLAYER.DELETE, target.getAttribute('id'));
+    }
+
+    /**
+     * Слушает клик мыши по кнопке лайка на треке в плейлисте
+     * @param {Object} event
+     */
+    trackFavoriteButtonClick(event) {
+        if (!User.exists()) {
+            globalEventBus.emit(GLOBAL.REDIRECT, URL.LOGIN);
+            return;
+        }
+        alert('This functionality is not accessible by now');
+    }
+
+    /**
+     * Слушает клик мыши по кнопке добавления на треке в плейлисте
+     * @param {Object} event
+     */
+    trackAddButtonClick(event) {
+        if (!User.exists()) {
+            globalEventBus.emit(GLOBAL.REDIRECT, URL.LOGIN);
+            return;
+        }
+        this._choosePlaylist.trackData = this.getIdByClick(event);
+        alert('NIKITA, I havent finished it yet');
+        this.getProfilePlaylists();
+    }
+
+    /**
+     * Получение id из dom-елемента по нажатию
+     * @param {Object} event
+     * @return {Object}
+     */
+    getIdByClick(event) {
+        let current = event.target;
+        while (current != null && !current.classList.contains('track-list')) {
+            if (current.classList.contains('border-bottom') &&
+                current.getAttribute('id') !== null) {
+                return {'id': current.getAttribute('id')};
+            }
+            current = current.parentNode;
+        }
+        return '';
+    }
+
+    /**
+     * Рисует треки в плейлисте
+     * @param {Object} tracks
+     */
+    drawTracklist(tracks) {
+        this.updateTrack(tracks[0]);
+        document.getElementsByClassName('track-list')[0].innerHTML += template(tracks);
+        this.locked = false;
+        if (!this.expanded) {
+            this.triggerClick();
+        }
+    }
+
+    /**
+     * Удаляет трек из списка воспроизвдения
+     * @param {string} id
+     */
+    removeFromTracklist(id) {
+        document.getElementById(id).remove();
+        if (document.getElementsByClassName('track-list')[0].children.length === 1) {
+            if (this.expanded) {
+                this.triggerClick();
+            }
+            this.locked = true;
+        }
+    }
+
+    /**
+     * Очищает список воспроизвдения
+     * @param {boolean} lock
+     */
+    removeFromTracklistAll(lock = false) {
+        while (document.getElementsByClassName('track-list')[0].children.length > 1) {
+            document.getElementsByClassName('track-list')[0].children[document
+                .getElementsByClassName('track-list')[0].children.length - 1].remove();
+        }
+        if (lock && this.expanded) {
+            this.triggerClick();
+        }
+        this.locked = lock;
+    }
+}
