+++ static/js/components/player_track_list_component/player_track_list_component.tmpl.xml
