+++ static/js/views/player/player.js
@@ -1,12 +1,11 @@
-import {PLAYER, NAVBAR, DOM, GLOBAL, URL} from '@libs/constans';
+import {PLAYER, NAVBAR, DOM, GLOBAL} from '@libs/constans';
 import BaseView from '@libs/base_view';
-import tracklist from '@views/player/tracklist.tmpl.xml';
 import player from '@views/player/player.tmpl.xml';
 import {globalEventBus} from '@libs/eventBus';
 import ChoosePlaylist from '@components/choose_playlist_component/choose_playlist_component';
-import User from '@libs/user';
 import PlaylistComponent from '@components/playlist_component/playlist_component';
 import TopTrackComponent from '@components/top_track_component/top_track_component';
+import TrackListComponent from '@components/player_track_list_component/player_track_list_component';
 
 /**
  *  вью для плеера
@@ -27,17 +26,23 @@ export default class PlayerView extends BaseView {
         this.muted = false;
         this.volume = 1;
         this.locked = true;
-        this.topTrackComponent = new TopTrackComponent();
+        this.topTrackComponent = new TopTrackComponent(eventBus);
+        this.trackListComponent = new TrackListComponent(eventBus);
         this._choosePlaylist = new ChoosePlaylist(eventBus, PLAYER);
-        this._playlistComponent = new PlaylistComponent(this.setDynamicEventListeners.bind(this));
+        this._playlistComponent = new PlaylistComponent(
+            this.trackListComponent.setEventListeners.bind(this));
+        this.eventBus.on(PLAYER.DRAW_TRACKLIST,
+            this.trackListComponent.drawTracklist.bind(this));
+        this.eventBus.on(PLAYER.DRAW_TRACKLIST,
+            this.trackListComponent.setEventListeners.bind(this.trackListComponent));
+        this.eventBus.on(PLAYER.REMOVE_FROM_TRACKLIST,
+            this.trackListComponent.removeFromTracklist.bind(this));
+        this.eventBus.on(PLAYER.REMOVE_FROM_TRACKLIST_ALL,
+            this.trackListComponent.removeFromTracklistAll.bind(this));
         this.eventBus.on(PLAYER.DRAW_PLAY, this.drawPlay.bind(this));
         this.eventBus.on(PLAYER.DRAW_PAUSE, this.drawPause.bind(this));
         this.eventBus.on(PLAYER.TRACK_UPDATE, this.updateTrack.bind(this));
         this.eventBus.on(PLAYER.MOVE_MARKER, this.moveMarker.bind(this));
-        this.eventBus.on(PLAYER.DRAW_TRACKLIST, this.drawTracklist.bind(this));
-        this.eventBus.on(PLAYER.DRAW_TRACKLIST, this.setDynamicEventListeners.bind(this));
-        this.eventBus.on(PLAYER.REMOVE_FROM_TRACKLIST, this.removeFromTracklist.bind(this));
-        this.eventBus.on(PLAYER.REMOVE_FROM_TRACKLIST_ALL, this.removeFromTracklistAll.bind(this));
         this.eventBus.on(PLAYER.DRAW_TIMELINE, this.drawTimeline.bind(this));
         this.eventBus.on(PLAYER.DRAW_SHUFFLE, this.drawShuffle.bind(this));
         this.eventBus.on(PLAYER.DRAW_UNSHUFLE, this.drawUnshuffle.bind(this));
@@ -57,7 +62,7 @@ export default class PlayerView extends BaseView {
     render(root, url) {
         super.render(document.getElementsByClassName(DOM.PLAYER)[0]);
         this.topTrackComponent.render();
-        this.setStaticEventListeners();
+        this.setEventListeners();
         this.resize();
     }
 
@@ -87,7 +92,7 @@ export default class PlayerView extends BaseView {
     /**
      * Sets static EventListeners
      */
-    setStaticEventListeners() {
+    setEventListeners() {
         window.addEventListener('resize', this.resize.bind(this));
         document.getElementsByTagName('body')[0]
             .addEventListener('DOMSubtreeModified', this.resize.bind(this));
@@ -170,25 +175,7 @@ export default class PlayerView extends BaseView {
             .onmousemove = (event) => this.volumeMouseMove(event);
         document.getElementsByClassName('volume-scale-front')[0]
             .onmousemove = (event) => this.volumeMouseMove(event);
-        window.onwheel = (event) => this.trackListWheel(event);
-    }
-
-    /**
-     * Sets dynamic EventListeners
-     */
-    setDynamicEventListeners() {
-        document.querySelectorAll('.track-list').forEach((row) => {
-            row.onclick = (event) => this.tracklistClick(event);
-        });
-        document.querySelectorAll('.delete-button').forEach((button) => {
-            button.onclick = (event) => this.trackDeleteButtonClick(event);
-        });
-        document.querySelectorAll('.favorite-button').forEach((button) => {
-            button.onclick = (event) => this.trackFavoriteButtonClick(event);
-        });
-        document.querySelectorAll('.add-button').forEach((button) => {
-            button.onclick = (event) => this.trackAddButtonClick(event);
-        });
+        window.onwheel = (event) => this.trackListComponent.trackListWheel(event);
     }
 
     /**
@@ -511,125 +498,6 @@ export default class PlayerView extends BaseView {
     }
 
     /**
-     * Слушает скрол для прокрутки плейлиста
-     * @param {Object} event
-     */
-    trackListWheel(event) {
-        const delta = event.deltaY;
-        const trackList = document.getElementsByClassName('track-list')[0];
-        if (event.clientX > trackList.getBoundingClientRect().x &&
-            event.clientX < trackList.getBoundingClientRect().x +
-            trackList.getBoundingClientRect().width &&
-            event.clientY > trackList.getBoundingClientRect().y &&
-            event.clientY < trackList.getBoundingClientRect().y +
-            trackList.getBoundingClientRect().height
-        ) {
-            event.preventDefault();
-            const top = parseInt(trackList.style.top.slice(0, trackList.style.top.length - 2));
-            if (delta < 0 && top < 0 ||
-                delta > 0 && trackList.getBoundingClientRect().bottom >
-                document.documentElement.clientHeight
-            ) {
-                if (delta < 0 && top - delta / 2 > 0) {
-                    trackList.style.top = '0';
-                } else if (delta > 0 && trackList.getBoundingClientRect().bottom - delta / 2 <
-                    document.documentElement.clientHeight
-                ) {
-                    const container = document.getElementsByClassName('container-audio')[0];
-                    trackList.style.top = (document.documentElement.clientHeight -
-                        trackList.getBoundingClientRect().height -
-                        container.getBoundingClientRect().bottom).toString() + 'px';
-                } else {
-                    trackList.style.top = (top - delta / 2).toString() + 'px';
-                }
-            }
-        }
-    }
-
-    /**
-     * Слушает клик по треку в плейлисте
-     * @param {Object} event
-     */
-    tracklistClick(event) {
-        let current = event.target;
-        if (current.classList.contains('m-obscure-title') ||
-            current.classList.contains('m-big-delete-button')) {
-            return;
-        }
-        while (current !== window && current !== document.body && current != null) { // TODO Сделать по-человечески :(
-            if (current.getAttribute('class') === 'track-list' ||
-                (current.getAttribute('class') !== null &&
-                    current.getAttribute('class').indexOf('button') !== -1 &&
-                    current.getAttribute('class').indexOf('buttons') === -1 &&
-                    current.getAttribute('class').indexOf('row') !== -1)
-            ) {
-                break;
-            }
-            if (current.getAttribute('id') !== null) {
-                this.eventBus.emit(PLAYER.GET_TRACK, current.getAttribute('id'));
-                break;
-            } else {
-                current = current.parentNode;
-            }
-        }
-    }
-
-    /**
-     * Слушает клик мыши по кнопке удаления на треке в плейлисте
-     * @param {Object} event
-     */
-    trackDeleteButtonClick(event) {
-        let target = event.target;
-        while (target.getAttribute('id') === null) {
-            target = target.parentNode;
-        }
-        this.eventBus.emit(PLAYER.DELETE, target.getAttribute('id'));
-    }
-
-    /**
-     * Слушает клик мыши по кнопке лайка на треке в плейлисте
-     * @param {Object} event
-     */
-    trackFavoriteButtonClick(event) {
-        if (!User.exists()) {
-            globalEventBus.emit(GLOBAL.REDIRECT, URL.LOGIN);
-            return;
-        }
-        alert('This functionality is not accessible by now');
-    }
-
-    /**
-     * Слушает клик мыши по кнопке добавления на треке в плейлисте
-     * @param {Object} event
-     */
-    trackAddButtonClick(event) {
-        if (!User.exists()) {
-            globalEventBus.emit(GLOBAL.REDIRECT, URL.LOGIN);
-            return;
-        }
-        this._choosePlaylist.trackData = this.getIdByClick(event);
-        alert('NIKITA, I havent finished it yet');
-        this.getProfilePlaylists();
-    }
-
-    /**
-     * Получение id из dom-елемента по нажатию
-     * @param {Object} event
-     * @return {Object}
-     */
-    getIdByClick(event) {
-        let current = event.target;
-        while (current != null && !current.classList.contains('track-list')) {
-            if (current.classList.contains('border-bottom') &&
-                current.getAttribute('id') !== null) {
-                return {'id': current.getAttribute('id')};
-            }
-            current = current.parentNode;
-        }
-        return '';
-    }
-
-    /**
      * Получение плейлистов пользователя
      */
     getProfilePlaylists() {
@@ -716,48 +584,6 @@ export default class PlayerView extends BaseView {
                 track1.style.marginLeft = '0';
             }
         }, 250);
-    }
-
-    /**
-     * Рисует треки в плейлисте
-     * @param {Object} tracks
-     */
-    drawTracklist(tracks) {
-        this.updateTrack(tracks[0]);
-        document.getElementsByClassName('track-list')[0].innerHTML += tracklist(tracks);
-        this.locked = false;
-        if (!this.expanded) {
-            this.triggerClick();
-        }
-    }
-
-    /**
-     * Удаляет трек из списка воспроизвдения
-     * @param {string} id
-     */
-    removeFromTracklist(id) {
-        document.getElementById(id).remove();
-        if (document.getElementsByClassName('track-list')[0].children.length === 1) {
-            if (this.expanded) {
-                this.triggerClick();
-            }
-            this.locked = true;
-        }
-    }
-
-    /**
-     * Очищает список воспроизвдения
-     * @param {boolean} lock
-     */
-    removeFromTracklistAll(lock = false) {
-        while (document.getElementsByClassName('track-list')[0].children.length > 1) {
-            document.getElementsByClassName('track-list')[0].children[document
-                .getElementsByClassName('track-list')[0].children.length - 1].remove();
-        }
-        if (lock && this.expanded) {
-            this.triggerClick();
-        }
-        this.locked = lock;
     }
 
     /**
